+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Added http rate limiter
+- Request timeout for ipfs fetch
+
 ## [0.1.3] - 2024-06-24
 
 ### Added